+++ CHANGELOG.md
@@ -9,3 +9,4 @@ HEAD
 * Add main README.md, LICENSE, and library.json files
 * Add main custom scripts for firmware packaging
 * Include custom scripts for firmware packaging cleanup
+* Add ESPInsights class source files